fixed TS compile error for session user id in leaving group

+++ src/components/utils/group/leaveGroup.tsx
@@ -8,23 +8,27 @@ export default function LeaveGroup() {
   const { supabase, session } = useSupabase();
 
   async function handleLeaveGroup() {
-    const { data, error } = await supabase
-      .from("groups")
-      .delete()
-      .eq("profile_id", session?.user.id);
+    if (session?.user?.id) {
+      const { data, error } = await supabase
+        .from("groups")
+        .delete()
+        .eq("profile_id", session?.user?.id);
 
-    if (error) {
-      throw new Error(error.message);
+      if (error) {
+        throw new Error(error.message);
+      } else {
+        console.log(data);
+        console.log("Group left!!");
+      }
     } else {
-      console.log(data);
-      console.log("Group left!!");
+      throw new Error("No session user Id found");
     }
   }
 
   return (
     <>
       <Button
-        sx={{ height: "fit-content",fontSize:"small" }}
+        sx={{ height: "fit-content", fontSize: "small" }}
         variant="contained"
         onClick={handleLeaveGroup}
         startIcon={<ExitToAppIcon />}
